message;

+++ saviors/WindowsFormsApp1/PhysX/IPlayer.cs
@@ -2,7 +2,7 @@
 {
     public interface IPlayer : IMovingBody
     {
-        byte Health { get; set; } 
+        int Health { get; set; } 
         
     }
 }
